backend/requirements.txt
openai
alpaca-py
python-dotenv
requests

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and validation
python-multipart==0.0.6